Read/write an image array to/from disk as text file(s).

Write images (arrays) as tab-separated .txt files on disk. Each channel-frame pair gets its own file.

write_txt_img(img, path, rds = FALSE, msg = TRUE)

read_txt_img(path, msg = TRUE)

Arguments

img
An image, represented by a 4-dimensional array, like an ijtiff_img.

path
The name of the input/output output file(s), without a file extension.

rds
In addition to writing a text file, save the image as an RDS (a single R object) file?

msg
Print an informative message about the image being read?

Examples

img <- read_tif(system.file('img', 'Rlogo.tif', package = 'ijtiff'))

#> Reading Rlogo.tif: an 8-bit, 76x100 pixel image of unsigned
#> integer type with 4 channels and 1 frame . . .

#>  Done.

tmptxt <- tempfile(pattern = "img", fileext = ".txt")
write_txt_img(img, tmptxt)

#> Writing imgc46cfcfbd45_ch1.txt, imgc46cfcfbd45_ch2.txt, imgc46cfcfbd45_ch3.txt and imgc46cfcfbd45_ch4.txt: a 76x100 pixel text image with 4 channels and 1 frame . . .

#>  Done.

tmptxt_ch1_path <- paste0(filesstrings::before_last_dot(tmptxt), "_ch1.txt")
print(tmptxt_ch1_path)

#> [1] "/var/folders/l_/2mwm03p55zg7zjykv084hhvr0000gn/T//RtmpBtCERe/imgc46cfcfbd45_ch1.txt"

txt_img <- read_txt_img(tmptxt_ch1_path)

#> Reading 76x100 pixel text image 'imgc46cfcfbd45_ch1.txt' . . .

#>  Done.