TIFF tag reference.

A dataset containing the information on all known baseline and extended TIFF tags.

tif_tags_reference()

Source

https://www.awaresystems.be

Details

A data frame with 96 rows and 10 variables:

code_dec

decimal numeric code of the TIFF tag

code_hex

hexadecimal numeric code of the TIFF tag

name

the name of the TIFF tag

short_description

a short description of the TIFF tag

tag_type

the type of TIFF tag: either "baseline" or "extended"

url

the URL of the TIFF tag at https://www.awaresystems.be

libtiff_name

the TIFF tag name in the libtiff C library

c_type

the C type of the TIFF tag data in libtiff

count

the number of elements in the TIFF tag data

default

the default value of the data held in the TIFF tag

Examples

tif_tags_reference()

#> # A tibble: 96 x 10
#>    code_dec code_hex name   short_description  type  url     libtiff_name c_type
#>       <dbl> <chr>    <chr>  <chr>              <chr> <chr>   <chr>        <chr> 
#>  1      254 00FE     NewSu… A general indicat… base… https:… TIFFTAG_SUB… LONG  
#>  2      255 00FF     Subfi… A general indicat… base… https:… TIFFTAG_OSU… SHORT 
#>  3      256 0100     Image… The number of col… base… https:… TIFFTAG_IMA… SHORT…
#>  4      257 0101     Image… The number of row… base… https:… TIFFTAG_IMA… SHORT…
#>  5      258 0102     BitsP… Number of bits pe… base… https:… TIFFTAG_BIT… SHORT 
#>  6      259 0103     Compr… Compression schem… base… https:… TIFFTAG_COM… SHORT 
#>  7      262 0106     Photo… The color space o… base… https:… TIFFTAG_PHO… SHORT 
#>  8      263 0107     Thres… For black and whi… base… https:… TIFFTAG_THR… SHORT 
#>  9      264 0108     CellW… The width of the … base… https:… TIFFTAG_CEL… SHORT 
#> 10      265 0109     CellL… The length of the… base… https:… TIFFTAG_CEL… SHORT 
#> # … with 86 more rows, and 2 more variables: count <chr>, default <chr>