Read TIFF tag information without actually reading the image array.

TIFF files contain metadata about images in their TIFF tags. This function is for reading this information without reading the actual image.

read_tags(path, all = TRUE)

Arguments

path
A string. The path to the tiff file to read.

all
TIFF files can contain multiple images. With all = TRUE, the information about all images is returned in a list of lists. To just get the information about some images, pass those image numbers to the all parameter (see examples). all = FALSE is equivalent to all = 1.

Value

A list of lists.

See also

Examples

read_tags(system.file("img", "Rlogo.tif", package="ijtiff"))

#> $img1
#> $img1$width
#> [1] 100
#> 
#> $img1$length
#> [1] 76
#> 
#> $img1$bits_per_sample
#> [1] 8
#> 
#> $img1$samples_per_pixel
#> [1] 4
#> 
#> $img1$sample_format
#> [1] "uint"
#> 
#> $img1$planar_config
#> [1] "contiguous"
#> 
#> $img1$rows_per_strip
#> [1] 76
#> 
#> $img1$compression
#> [1] "LZW"
#> 
#> $img1$x_resolution
#> [1] 299.99
#> 
#> $img1$y_resolution
#> [1] 299.99
#> 
#> $img1$resolution_unit
#> [1] "inch"
#> 
#> $img1$orientation
#> [1] "top_left"
#> 
#> $img1$color_space
#> [1] "RGB"
#> 
#> 

read_tags(system.file("img", "2ch_ij.tif", package="ijtiff"))

#> $img1
#> $img1$width
#> [1] 6
#> 
#> $img1$length
#> [1] 15
#> 
#> $img1$bits_per_sample
#> [1] 8
#> 
#> $img1$samples_per_pixel
#> [1] 1
#> 
#> $img1$sample_format
#> [1] "uint"
#> 
#> $img1$planar_config
#> [1] "contiguous"
#> 
#> $img1$rows_per_strip
#> [1] 15
#> 
#> $img1$compression
#> [1] "none"
#> 
#> $img1$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img1$color_space
#> [1] "black is zero"
#> 
#> 
#> $img2
#> $img2$width
#> [1] 6
#> 
#> $img2$length
#> [1] 15
#> 
#> $img2$bits_per_sample
#> [1] 8
#> 
#> $img2$samples_per_pixel
#> [1] 1
#> 
#> $img2$sample_format
#> [1] "uint"
#> 
#> $img2$planar_config
#> [1] "contiguous"
#> 
#> $img2$rows_per_strip
#> [1] 15
#> 
#> $img2$compression
#> [1] "none"
#> 
#> $img2$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img2$color_space
#> [1] "black is zero"
#> 
#> 
#> $img3
#> $img3$width
#> [1] 6
#> 
#> $img3$length
#> [1] 15
#> 
#> $img3$bits_per_sample
#> [1] 8
#> 
#> $img3$samples_per_pixel
#> [1] 1
#> 
#> $img3$sample_format
#> [1] "uint"
#> 
#> $img3$planar_config
#> [1] "contiguous"
#> 
#> $img3$rows_per_strip
#> [1] 15
#> 
#> $img3$compression
#> [1] "none"
#> 
#> $img3$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img3$color_space
#> [1] "black is zero"
#> 
#> 
#> $img4
#> $img4$width
#> [1] 6
#> 
#> $img4$length
#> [1] 15
#> 
#> $img4$bits_per_sample
#> [1] 8
#> 
#> $img4$samples_per_pixel
#> [1] 1
#> 
#> $img4$sample_format
#> [1] "uint"
#> 
#> $img4$planar_config
#> [1] "contiguous"
#> 
#> $img4$rows_per_strip
#> [1] 15
#> 
#> $img4$compression
#> [1] "none"
#> 
#> $img4$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img4$color_space
#> [1] "black is zero"
#> 
#> 
#> $img5
#> $img5$width
#> [1] 6
#> 
#> $img5$length
#> [1] 15
#> 
#> $img5$bits_per_sample
#> [1] 8
#> 
#> $img5$samples_per_pixel
#> [1] 1
#> 
#> $img5$sample_format
#> [1] "uint"
#> 
#> $img5$planar_config
#> [1] "contiguous"
#> 
#> $img5$rows_per_strip
#> [1] 15
#> 
#> $img5$compression
#> [1] "none"
#> 
#> $img5$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img5$color_space
#> [1] "black is zero"
#> 
#> 
#> $img6
#> $img6$width
#> [1] 6
#> 
#> $img6$length
#> [1] 15
#> 
#> $img6$bits_per_sample
#> [1] 8
#> 
#> $img6$samples_per_pixel
#> [1] 1
#> 
#> $img6$sample_format
#> [1] "uint"
#> 
#> $img6$planar_config
#> [1] "contiguous"
#> 
#> $img6$rows_per_strip
#> [1] 15
#> 
#> $img6$compression
#> [1] "none"
#> 
#> $img6$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img6$color_space
#> [1] "black is zero"
#> 
#> 
#> $img7
#> $img7$width
#> [1] 6
#> 
#> $img7$length
#> [1] 15
#> 
#> $img7$bits_per_sample
#> [1] 8
#> 
#> $img7$samples_per_pixel
#> [1] 1
#> 
#> $img7$sample_format
#> [1] "uint"
#> 
#> $img7$planar_config
#> [1] "contiguous"
#> 
#> $img7$rows_per_strip
#> [1] 15
#> 
#> $img7$compression
#> [1] "none"
#> 
#> $img7$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img7$color_space
#> [1] "black is zero"
#> 
#> 
#> $img8
#> $img8$width
#> [1] 6
#> 
#> $img8$length
#> [1] 15
#> 
#> $img8$bits_per_sample
#> [1] 8
#> 
#> $img8$samples_per_pixel
#> [1] 1
#> 
#> $img8$sample_format
#> [1] "uint"
#> 
#> $img8$planar_config
#> [1] "contiguous"
#> 
#> $img8$rows_per_strip
#> [1] 15
#> 
#> $img8$compression
#> [1] "none"
#> 
#> $img8$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img8$color_space
#> [1] "black is zero"
#> 
#> 
#> $img9
#> $img9$width
#> [1] 6
#> 
#> $img9$length
#> [1] 15
#> 
#> $img9$bits_per_sample
#> [1] 8
#> 
#> $img9$samples_per_pixel
#> [1] 1
#> 
#> $img9$sample_format
#> [1] "uint"
#> 
#> $img9$planar_config
#> [1] "contiguous"
#> 
#> $img9$rows_per_strip
#> [1] 15
#> 
#> $img9$compression
#> [1] "none"
#> 
#> $img9$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img9$color_space
#> [1] "black is zero"
#> 
#> 
#> $img10
#> $img10$width
#> [1] 6
#> 
#> $img10$length
#> [1] 15
#> 
#> $img10$bits_per_sample
#> [1] 8
#> 
#> $img10$samples_per_pixel
#> [1] 1
#> 
#> $img10$sample_format
#> [1] "uint"
#> 
#> $img10$planar_config
#> [1] "contiguous"
#> 
#> $img10$rows_per_strip
#> [1] 15
#> 
#> $img10$compression
#> [1] "none"
#> 
#> $img10$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img10$color_space
#> [1] "black is zero"
#> 
#> 

read_tags(system.file("img", "2ch_ij.tif", package="ijtiff"), all = c(2, 4))

#> $img2
#> $img2$width
#> [1] 6
#> 
#> $img2$length
#> [1] 15
#> 
#> $img2$bits_per_sample
#> [1] 8
#> 
#> $img2$samples_per_pixel
#> [1] 1
#> 
#> $img2$sample_format
#> [1] "uint"
#> 
#> $img2$planar_config
#> [1] "contiguous"
#> 
#> $img2$rows_per_strip
#> [1] 15
#> 
#> $img2$compression
#> [1] "none"
#> 
#> $img2$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img2$color_space
#> [1] "black is zero"
#> 
#> 
#> $img4
#> $img4$width
#> [1] 6
#> 
#> $img4$length
#> [1] 15
#> 
#> $img4$bits_per_sample
#> [1] 8
#> 
#> $img4$samples_per_pixel
#> [1] 1
#> 
#> $img4$sample_format
#> [1] "uint"
#> 
#> $img4$planar_config
#> [1] "contiguous"
#> 
#> $img4$rows_per_strip
#> [1] 15
#> 
#> $img4$compression
#> [1] "none"
#> 
#> $img4$description
#> [1] "ImageJ=1.51s\nimages=10\nchannels=2\nframes=5\nhyperstack=true\nmode=composite\nloop=false\n"
#> 
#> $img4$color_space
#> [1] "black is zero"
#> 
#>